Don’t specify group when creating archivematica user
If the group specified doesn’t yet exist, this task will error out.
However, if no group is specified, then a group will be created with
the same name as the user - which is the desired effect anyway.

This currently only works because the user is being implicitly
created by the “postinst” script run as a part of the storage
service; as a result, the user already exists at every point the
“create user archivematica” task is called.