Don't run freshclam on installation
freshclam stalls often (probably caused by clamav
server issues) when run by this role.
As a workaround, download the clamav signatures
from a github repo to avoid the need to run freshclam.
The signature files are recent but not guaranteed
to be the latest ones, however those should be
updated by freshclam later.