Configure AM17 logging
A new option to configure backward-compatible logging has been added to
substitute the new AM17 logging setup. The default AM17 logging sends
the events to the standard streams, which is more convenient when
Archivematica is running in a cluster.

In order to use the backward-compatible logging, the boolean environment
variable 'archivematica_src_logging_backward_compatible' has to be
enabled, which is the default behaviour.

The log directory paths and log file sizes are configurable for each
service.