pipeline-websrv: Fix remove default nginx site
This commit changes the task "Remove Nginx default server", deleting the
enabled default site instead of the available one. When deploying AM
pipeline only, the nginx fails to start because exists the enabled
directory default file but not the available directory one.

When deploying pipelines and ss this issue does not happen because the
enabled default file was deleted by the ss-websrv "Remove Nginx default
server" task.