pipeline-es: Add ES index mappings workaround
This commit adds the workaround for issue:

https://github.com/archivematica/Issues/issues/608

It is is needed when upgrading from AM v1.9.0.

The archivematica_src_search_enabled variable can be set as:

- true
- yes
- all
- false
- aips
- transfers

That's the reason why aips and transfers indices are checked
individually and there are 2 separate tasks to add the workaround for
transfers and aips.