configure-gpg: Change the way gpg key is created
The gpg key is created with python's gpg module instead of using gpg
command directly. This is the same way that key is created by the
Archivematica Storage Service.

The `binary` key was removed from `archivematica_src_configure_gpg`
dict, now it is defined in a tasks in role (Bionic=gpg1, Others=gpg)