Add become_flag=-E to git clone tasks
The ansible role, when was run with `become=yes` and `ForwardAgent=yes`
options, was using the `ansible_host` ssh keys for the `ansible_ssh_user` user,
but not for the root user that was running the git clone command. So the git
clone tasks were failing when using the git transport protocol and `become=yes`
option.

Using the `-E` become_flag allows to use the `ansible_ssh_user` environment in
the git clone tasks, so the `ansible_ssh_user`'s `SSH_AUTH_SOCK` variable is
used by the root user.